backend/CV backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain==0.1.1
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
